apply suggestions and test chunking itself
