+++ packages/static-build/test/integration-3.test.js
@@ -1,1 +1,1 @@
-require('./integration-1.test');
+require('./integration-setup')(3);
