+++ packages/static-build/test/integration-5.test.js
@@ -1,1 +1,1 @@
-require('./integration-1.test');
+require('./integration-setup')(5);
